Added IRQ_Register
added the irq_register that was missing, managed to get everything
compiling not sure how the behaviour is as extensions to the testbench
will have to be made.. logic for the data calls not completed yet. used
the manual to determine most of what was done on page 7 at the bottom.